Refactor: удалить дублирующую валидацию после ревью
