+++ src/main/java/ru/yandex/practicum/filmorate/storage/dao/ReviewDbStorage.java
@@ -32,7 +32,7 @@ public class ReviewDbStorage implements ReviewStorage {
 
     @Override
     public Review addNewReview(Review review) {
-        idValidation(review);
+
         String sqlQueryReview = "insert into reviews(content, is_positive, user_id, film_id) " +
                 "values (?, ?, ?, ?) ";
         KeyHolder keyHolder = new GeneratedKeyHolder();           // вернуть id, сгенерированный в БД
@@ -53,7 +53,6 @@ public class ReviewDbStorage implements ReviewStorage {
 
     @Override
     public Review update(Review review) {
-        getReviewById(review.getReviewId());
         String sqlQuery = "UPDATE reviews SET " +
                 "content = ?, is_positive = ? " +
                 "WHERE review_id = ? ";
@@ -66,7 +65,6 @@ public class ReviewDbStorage implements ReviewStorage {
 
     @Override
     public void delete(Integer id) {
-        getReviewById(id);
         String sqlQueryDeleteReview = "DELETE FROM reviews WHERE review_id = ? ";
         jdbcTemplate.update(sqlQueryDeleteReview, id);
         log.info("Отзыв с идентификатором " + id + " удалён из базы");
@@ -118,14 +116,5 @@ public class ReviewDbStorage implements ReviewStorage {
                 rs.getInt("user_id"),
                 rs.getInt("film_id"),
                 rs.getInt("useful"));
-    }
-
-    private void idValidation(Review review) {
-        if (review.getUserId() < 1) {
-            throw new UserNotFoundException("id пользователя не может быть меньше 1");
-        }
-        if (review.getFilmId() < 1) {
-            throw new FilmNotFoundException("id фильма не может быть меньше 1");
-        }
     }
 }
